+++ src/Supports/Timber.php
@@ -10,8 +10,10 @@ declare(strict_types=1);
 
 namespace Picowind\Supports;
 
+use GuzzleHttp\Utils;
 use Picowind\Core\Discovery\Attributes\Hook;
 use Picowind\Core\Discovery\Attributes\Service;
+use Picowind\Utils\Theme as UtilsTheme;
 use Timber\Site;
 use Timber\Timber as TimberTimber;
 
@@ -87,7 +89,9 @@ class Timber extends Site
         // If it didn't (meaning Twig rendering failed and no fallback was triggered),
         // then include index.php as fallback
         if (false === $output || empty(trim($output))) {
-            include get_template_directory() . '/index.php';
+            include UtilsTheme::is_child_theme() && file_exists(UtilsTheme::child_dir() . '/index.php')
+                ? UtilsTheme::child_dir() . '/index.php'
+                : UtilsTheme::parent_dir() . '/index.php';
         } else {
             echo $output;
         }